backend/pyproject.toml
[project]
name = "erp-metalfort-backend"
version = "0.1.0"
description = "FastAPI backend for ERP Metalfort"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic[email]>=2.9",
    "pydantic-settings>=2.5",
    "httpx>=0.27",
    "supabase>=2.8",
    "pyjwt[crypto]>=2.9",
    "python-multipart>=0.0.12",
    "slowapi>=0.1.9",
    "weasyprint>=62",
    "resend>=2.4",
    "jinja2>=3.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
